Simplifies build process
Before, the TS source was actually transpiled three times - once for the node build with tsc, and twice for the browser bundles with ts-loader. This was extremely slow, taking two seconds per bundle with webpack.

Now, the TS is transpiled *once*, and webpack bundles the output in lib/. This speeds up builds immensely.

Also adds a new webpack plugin to preserve the original sourcemaps across multiple transformations.